[#28] storagegroup: Measure lifetime in NeoFS epochs
In previous version of the format storage group lifetime was measured in two
epoch times: NeoFS and Unix. Since unix is not processed in the system,
this commit narrows the time point of storage group expiration to the
system epoch number. The storagegroup.StorageGroup.Lifetime message has
therefore been deleted as no longer used.

Signed-off-by: Leonard Lyubich <leonard@nspcc.ru>