refs: Define ContainerID message
This commit defines ContainerID message in refs package. In the future, this
structure can be ported into messages in place of the container identifier
field to maintain format uniformity. Also, storing the identifier in a
dedicated message will allow, if necessary, to expand it with additional
information without losing backward compatibility within one version of the
API.

Signed-off-by: Leonard Lyubich <leonard@nspcc.ru>