[#31] service: Change BearerTokenMsg.Info.ACLRules field type
It was originally planned to transport extended ACL table in binary form.
However, from time on, it was decided to abandon this approach in favor of
the declared message structure (acl.EACLTable). In this regard, this commit
changes the type and name of the binary BearerTokenMsg.Info.ACLRules field
to EACLTable.

Signed-off-by: Leonard Lyubich <leonard@nspcc.ru>