[#202] session: Support sessions issued for an object group
Sometimes it is useful to open single session which covers several
objects in the container. In previous implementation it could be done:
 * by opening container-global session
 * by opening N per-object sessions

Both approaches are not optimal for the mentioned need.

Define `ObjectSessionContext.Target` which is backward compatible with
`refs.Address` in binary format. Replace `address` field of
`refs.Address` type in `ObjectSessionContext` message with `target`
field of `ObjectSessionContext.Target` type. This change saves backward
compatibility in binary format, but break the JSON one. Such a breakage
is considered admissible for now since NeoFS API protocol hasn't
declared JSON compatibility yet.

Signed-off-by: Leonard Lyubich <leonard@nspcc.ru>