netmap: Add reserved node attribute for verified nodes' domain
Sometimes there is a need to control the entry of some storage nodes
into a certain group at the stage of registration in the NeoFS network.
For example, a specific organization may wish to be part of a public
network, but limit the ability of a third party to indicate its
affiliation. Previously, any candidate node could specify any valid
attribute making private node group (aka subnet) not implementable.

This commit reserves node attribute with key `VerifiedNodesDomain`. From
now, any storage attempting to enter the NeoFS network and declaring
this attribute will have to confirm its right to do so. This feature
will allow parties to organize subgroups of nodes (i.e. subnets) and
control access to them.

While new changes impose requirements and are technically not backwards
compatible, the use of the new attribute has never been observed in
practice. It's worth noting that omitting is backwards compatible.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>