acl: Support access rules for absent attributes
Previously, protocol did not clearly support access rules based on
attribute absence. Using `STRING_EQUAL` op with empty value was an
insufficiently clear notation of a missing attribute since it could
cause a collision with attributes with an empty value (there are no such
things now, but they may be possible in the future).

`MatchType` enumeration is extended with `NOT_PRESENT` value. Being set
in the `EACLRecord.Filter`, this operator will limit access rule to the
objects without. The op is prohibited for system attributes so as not to
create deliberately false (for known attributes) or undefined behavior
(for unsupported ones).

Closes #256.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>