tombstone: Require members to be root objects
It simplifies (and allows ACL problems solution in some cases) client
implementations and makes it clearer for users. Any split information should be
accessible on the node side, otherwise it is also inaccessible on the client
side.
Garbage collection after incomplete PUTs MUST be allowed anyway, so LINK object
absence is an exception to the new rule.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>