refs: Prohibit IDs consisting of all zero bytes
Container and object IDs are SHA-256 checksums of their contents. It is
almost impossible to find data that corresponds to a zero identifier.
Setting 32 zeros rather indicates a client error. Reserving a zero value
will also make application development easier: corresponding type will
have a reserved invalid value within itself. Even if the unimaginable
and some structure will actually have a zero hash - so be it, the
problem is negligible for the final benefit.

Note that zero `OwnerID` is already invalid at least due to the non-zero
prefix. Therefore does not need clarification.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>