object: Allow header with payload chunk in first GET response message
Previously, heading and chunk fields were mutually exclusive in each GET
response message. With this, server sent 1st message with the header
only, then messages with payload chunks came. This behavior is
inefficient since any object with a non-empty payload requires one more
message. The smaller the payload, the greater the overhead.

This modifies GET protocol to allow header and chunk to be attached to
the same message. For backward compatibility, requests from older
clients will be processed as before.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>